Added colcon to build workspace and source it
This commit introduces a colcon builder to build a colcon workspace using
robot_folders. For it to work, the colcon workspace has to be in a folder
called 'colcon_ws' inside the environment. It cannot be created by the setup
assistant, create it by hand currently.

To build a colcon workspace, use either of
 - fzirob make colcon
 - makecol

It should be sourced together with the environment, but I'm not sure, whether
it works correctly.