Made completion work again on zsh
Since we install using pipx, we install click in the virtualenv in
version 8.x. However, if the host system has installed click in a lower
version (or not installed at all) this will fail if we use the standard
python interpreter to check for the click version.

The change from this commit uses the script from the PATH instead of
using an absolute path together with the standard interpreter, which in
turn uses the click version installed in the virtualenv (or from the
system if installed otherwise).