Allow chaining of workspaces
Though it is not recommended to chain workspaces it might be desirable to
source a workspace that us not the currently sourced one. With the changes
introduced in version 0.5.0 this possibility was removed.
This commit re-adds that functionality. The .cur_env file is always read
independent of whether the ROB_FOLDERS_ACTIVE_ENV variable is set or not.